third_party/WebKit/LayoutTests/fast/canvas/canvas-filter-value-expected.txt

PASS ctx.filter is "none"
PASS ctx.filter is "blur(5px)"
PASS ctx.filter is "none"
PASS ctx.filter is "blur(5px)"
PASS ctx.filter is "blur(5px)"
PASS ctx.filter is "blur(5px)"
PASS ctx.filter is "blur(5px)"
PASS ctx.filter is "blur(5px)"
PASS ctx.filter is "blur(5px)"
PASS ctx.filter is "blur(5px)"
PASS ctx.filter is "blur(5px)"
PASS ctx.filter is "blur(  5px)"
PASS successfullyParsed is true

TEST COMPLETE
